+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v53.2.0...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v54.0.0...main)
+
+# v54.0.0 (2023-09-12)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v53.2.0...v54.0.0)
+
+* General
+  * Experimental: Add configuration to add a wall clock timestamp to all events ([#2513](https://github.com/mozilla/glean/issues/2513))
+* Python
+  * Switched the build system to maturin. This should not have any effect on consumers. ([#2345](https://github.com/mozilla/glean/pull/2345))
+  * BREAKING CHANGE: Dropped support for Python 3.6 ([#2345](https://github.com/mozilla/glean/pull/2345))
+* Kotlin
+  * Update to Gradle v8.2.1 ([#2516](https://github.com/mozilla/glean/pull/2516))
+  * Increase Android compile SDK to version 34 ([#2614](https://github.com/mozilla/glean/pull/2614))
 
 # v53.2.0 (2023-08-02)
 
@@ -11,7 +24,7 @@
 * Rust
   * The Ping Rate Limit type is now accessible in the Rust Language Binding ([#2528](https://github.com/mozilla/glean/pull/2528))
   * Gracefully handle a failure when starting the upload thread. Glean no longer crashes in that case. ([#2545](https://github.com/mozilla/glean/pull/2545))
-  * `locale` now exposed through the RLB so it can be set by consumers ([2531](https://github.com/mozilla/glean/pull/2531))
+  * `locale` now exposed through the RLB so it can be set by consumers ([#2531](https://github.com/mozilla/glean/pull/2531))
 * Python
   * Added the shutdown API for Python to ensure orderly shutdown and waiting for uploader processes ([#2538](https://github.com/mozilla/glean/pull/2538))
 * Kotlin
@@ -262,9 +275,6 @@
 
 * Kotlin
   * Update Kotlin and Android Gradle Plugin to the latest releases ([#2211](https://github.com/mozilla/glean/pull/2211))
-* Swift
-  * Fix for iOS startup crash caused by Glean ([#2206](https://github.com/mozilla/glean/pull/2206))
-
 * Swift
   * Fix for iOS startup crash caused by Glean ([#2206](https://github.com/mozilla/glean/pull/2206))